Refactor and generalize code
