+++ extension.json
@@ -60,7 +60,7 @@
     "WSSlotsDoPurge": {
       "value": false
     },
-    "WSSlotsOverrideRawAction": {
+    "WSSlotsOverrideActions": {
       "value": false
     }
   },
@@ -70,8 +70,8 @@
     "readslot": "WSSlots\\API\\ApiReadSlot"
   },
   "Actions": {
-		"rawslot": "SlotAwareRawAction"
-	},
+		"rawslot": "WSSlots\\Actions\\SlotAwareRawAction"
+  },
   "manifest_version": 2,
   "load_composer_autoloader": true
 }
